simplify substate transformer
