+++ lib/built_redux.dart
@@ -3,7 +3,6 @@ library built_redux;
 export 'src/action.dart';
 export 'src/reducer_builder.dart';
 export 'src/middleware.dart';
-export 'src/state_transformer.dart';
 export 'src/store.dart';
 export 'src/store_change.dart';
 export 'src/typedefs.dart';
